Move HTTPPollerSource dependency under io.cdap.plugin.spark
