+++ spark-plugins/src/main/java/io/cdap/plugin/spark/HTTPPollerSource.java
@@ -22,8 +22,6 @@ import io.cdap.cdap.api.annotation.Plugin;
 import io.cdap.cdap.api.data.format.StructuredRecord;
 import io.cdap.cdap.etl.api.streaming.StreamingContext;
 import io.cdap.cdap.etl.api.streaming.StreamingSource;
-import io.cdap.plugin.common.http.HTTPPollConfig;
-import io.cdap.plugin.common.http.HTTPRequestor;
 import org.apache.spark.storage.StorageLevel;
 import org.apache.spark.streaming.api.java.JavaDStream;
 import org.apache.spark.streaming.receiver.Receiver;
